pyproject.toml
# pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "my-ai-assistant"
version = "1.0.0"
description = "AI Assistant for Software Development"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "requests",
    "pyyaml",
    "pydantic>=2.0"
]

[project.scripts]
ai = "ai_assistant.cli:main"

[tool.setuptools.package-data]
ai_assistant = ["personas/**/*.md"]